refactor: create deno_lib crate
