+++ Cargo.toml
@@ -28,6 +28,7 @@ members = [
   "ext/webidl",
   "ext/websocket",
   "ext/webstorage",
+  "lib",
   "resolvers/deno",
   "resolvers/node",
   "resolvers/npm_cache",
@@ -93,7 +94,8 @@ deno_webidl = { version = "0.185.0", path = "./ext/webidl" }
 deno_websocket = { version = "0.190.0", path = "./ext/websocket" }
 deno_webstorage = { version = "0.180.0", path = "./ext/webstorage" }
 
-# resolvers
+# workspace libraries
+deno_lib = { version = "=0.1.1", path = "./lib" }
 deno_npm_cache = { version = "0.4.0", path = "./resolvers/npm_cache" }
 deno_resolver = { version = "0.16.0", path = "./resolvers/deno" }
 node_resolver = { version = "0.23.0", path = "./resolvers/node" }
