+++ cli/cache/emit.rs
@@ -6,19 +6,20 @@ use deno_ast::ModuleSpecifier;
 use deno_core::anyhow::anyhow;
 use deno_core::error::AnyError;
 use deno_core::unsync::sync::AtomicFlag;
+use deno_lib::cache::DiskCache;
 
-use super::DiskCache;
+use crate::sys::CliSys;
 
 /// The cache that stores previously emitted files.
 #[derive(Debug)]
 pub struct EmitCache {
-  disk_cache: DiskCache,
+  disk_cache: DiskCache<CliSys>,
   emit_failed_flag: AtomicFlag,
   file_serializer: EmitFileSerializer,
 }
 
 impl EmitCache {
-  pub fn new(disk_cache: DiskCache) -> Self {
+  pub fn new(disk_cache: DiskCache<CliSys>) -> Self {
     Self {
       disk_cache,
       emit_failed_flag: Default::default(),
