+++ lib/cache/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2018-2025 the Deno authors. MIT license.
+
+pub use deno_dir::DenoDir;
+pub use deno_dir::DenoDirProvider;
+pub use disk_cache::DiskCache;
+
+mod deno_dir;
+mod disk_cache;
