+++ lib/lib.rs
@@ -1,0 +1,6 @@
+// Copyright 2018-2025 the Deno authors. MIT license.
+
+pub mod cache;
+pub mod npm;
+pub mod standalone;
+pub mod sys;
