+++ lib/standalone/mod.rs
@@ -1,0 +1,3 @@
+// Copyright 2018-2025 the Deno authors. MIT license.
+
+pub mod virtual_fs;
